backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.0
python-multipart==0.0.21  # Required for form parsing in Starlette

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0

# GraphQL
strawberry-graphql[asgi]==0.220.0

# AWS/S3
boto3==1.34.14
botocore==1.34.14
boto3-stubs[s3]==1.34.14  # Type stubs for boto3

# Development dependencies (optional)
# Install with: pip install -r requirements.txt -r requirements-dev.txt
# Or: pip install ".[dev]"
